Add option to String::assign() to build a string from two char blocks.
Use the modified String::assign() method to implement a contructor for NamedString which can put a prefix to a name.
Add to the NamedList::addParam() methods the option to pass a prefix for the names.

This allows eliminating the necessity of building prefixed named before passing them to NamedList::addParam().